feat(analytics): refactoring middleware
