+++ packages/blockchain-wallet-v4-frontend/src/middleware/analyticsMiddleware/index.ts
@@ -1,250 +1,20 @@
-import { v4 } from 'uuid'
+import analytics from 'middleware/analyticsMiddleware/analytics'
+import type { PageNamesType } from 'middleware/analyticsMiddleware/types'
+import { AnalyticsKey, AnalyticsType } from 'middleware/analyticsMiddleware/types'
+import {
+  getOriginalTimestamp,
+  simpleBuyOriginDictionary,
+  simpleBuyPaymentTypeDictionary
+} from 'middleware/analyticsMiddleware/utils'
 
 import { getCardTypeByValue } from 'components/Form/CreditCardBox/model'
-import { SBPaymentTypes } from 'core/types'
 import { actionTypes as AT } from 'data'
-import { ModalNamesType, SBShowModalOriginType } from 'data/types'
-import queuevent from 'utils/queuevent'
-
-enum AnalyticsKey {
-  BUY_AMOUNT_ENTERED = 'Buy Amount Entered',
-  BUY_AMOUNT_MAX_CLICKED = 'Buy Amount Max Clicked',
-  BUY_AMOUNT_MIN_CLICKED = 'Buy Amount Min Clicked',
-  // BUY_LEARN_MORE_CLICKED = 'Buy Learn More Clicked', // not implemented
-  BUY_PAYMENT_METHOD_SELECTED = 'Buy Payment Method Selected',
-  BUY_SELL_CLICKED = 'Buy Sell Clicked',
-  BUY_SELL_VIEWED = 'Buy Sell Viewed',
-  CARD_REJECTED = 'Card Rejected',
-  DASHBOARD_CLICKED = 'Dashboard Clicked',
-  DASHBOARD_VIEWED = 'Dashboard Viewed',
-  EMAIL_VERIFICATION_REQUESTED = 'Email Verification Requested',
-  SIGNED_IN = 'Signed In',
-  SIGNED_OUT = 'Signed Out',
-  // SWAP_ACCOUNTS_SELECTED = 'Swap Accounts Selected', // not implemented
-  // SWAP_AMOUNT_ENTERED = 'Swap Amount Entered', // not implemented
-  // SWAP_AMOUNT_MAX_CLICKED = 'Swap Amount Max Clicked', // not implemented
-  // SWAP_AMOUNT_MIN_CLICKED = 'Swap Amount Min Clicked', // not implemented
-  SWAP_CLICKED = 'Swap Clicked',
-  // SWAP_FROM_SELECTED = 'Swap From Selected', // not implemented
-  // SWAP_RECEIVE_SELECTED = 'Swap Receive Selected', // not implemented
-  SWAP_VIEWED = 'Swap Viewed',
-  UPGRADE_VERIFICATION_CLICKED = 'Upgrade Verification Clicked'
-}
-
-enum AnalyticsType {
-  EVENT = 'EVENT',
-  VIEW = 'VIEW'
-}
-
-type BasePayload = {
-  id?: string
-  originalTimestamp: string
-  type: AnalyticsType
-}
-
-type PageViewPayload = {
-  path: string
-  referrer: string
-  search: string
-  title: string
-  url: string
-}
-
-type BuyAmountEnteredPayload = BasePayload & {
-  input_amount: number
-  input_currency: string
-  max_card_limit: number
-  output_currency: string
-  platform: 'WALLET'
-}
-
-type BuyAmountMaxClickedPayload = BasePayload & {
-  input_currency: string
-  max_card_limit: number
-  output_currency: string
-  platform: 'WALLET'
-}
-
-type BuyAmountMinClickedPayload = BasePayload & {
-  input_currency: string
-  output_currency: string
-  platform: 'WALLET'
-}
-
-type BuyLearnMoreClickedPayload = BasePayload & {
-  platform: 'WALLET'
-}
-
-type BuyPaymentMethodSelectedPayload = BasePayload & {
-  payment_type: 'BANK_ACCOUNT' | 'BANK_TRANSFER' | 'FUNDS' | 'PAYMENT_CARD'
-  platform: 'WALLET'
-}
-
-type BuySellClickedPayload = BasePayload & {
-  origin: 'BUY_WIDGET' | string
-  // type: "BUY" | "SELL"
-}
-
-type BuySellViewedPayload = BasePayload &
-  PageViewPayload & {
-    // type: "BUY" | "SELL"
-  }
-
-type CardRejectedPayload = BasePayload & {
-  card_type: string
-  country_billing: string
-  product: 'SIMPLE_BUY' | 'SIMPLE_TRADE' | 'SWAP'
-  reason: string
-}
-
-type SwapClickedPayload = BasePayload & {
-  origin: 'CURRENCY_PAGE' | 'DASHBOARD_PROMO' | 'NAVIGATION'
-}
-
-type SwapViewedPayload = BasePayload & PageViewPayload & {}
-
-type DashboardClickedPayload = BasePayload & {
-  origin: 'SIGN_IN'
-}
-
-type DashboardViewedPayload = BasePayload & PageViewPayload & {}
-
-type EmailVerificationClicked = BasePayload & {
-  // origin: 'SIGN_UP' | 'VERIFICATION'
-}
-
-type SignedInPayload = BasePayload & {
-  platform: 'WALLET'
-}
-
-type SignedUpPayload = BasePayload & {
-  platform: 'WALLET'
-}
-
-type UpgradeVerificationClickedPayload = BasePayload & {
-  // origin:
-  //   | 'AIRDROP'
-  //   | 'FIAT_FUNDS'
-  //   | 'RESUBMISSION'
-  //   | 'SAVINGS'
-  //   | 'SETTINGS'
-  //   | 'SIMPLEBUY'
-  //   | 'SIMPLETRADE'
-  //   | 'SWAP'
-  tier: number
-}
-
-type AnalyticsPayload =
-  | BuyAmountEnteredPayload
-  | BuyAmountMaxClickedPayload
-  | BuyAmountMinClickedPayload
-  | BuyLearnMoreClickedPayload
-  | BuyPaymentMethodSelectedPayload
-  | BuySellClickedPayload
-  | BuySellViewedPayload
-  | CardRejectedPayload
-  | SwapClickedPayload
-  | SwapViewedPayload
-  | DashboardClickedPayload
-  | DashboardViewedPayload
-  | EmailVerificationClicked
-  | SignedInPayload
-  | SignedUpPayload
-  | UpgradeVerificationClickedPayload
-
-type PageNamesType = '/home'
-// | '/interest'
-// | '/borrow'
-// | '/settings/general'
-// | '/settings/preferences'
-// | '/settings/addresses'
-
-const simpleBuyOriginDictionary = (rawOrigin: SBShowModalOriginType) => {
-  switch (rawOrigin) {
-    case 'InterestPage':
-      return 'SAVINGS'
-    case 'PendingOrder':
-      return 'PENDING_ORDER'
-    case 'SideNav':
-      return 'NAVIGATION'
-    case 'WelcomeModal':
-      return 'WELCOME'
-    case 'PriceChart':
-      return 'PRICE_CHART'
-    case 'SimpleBuyLink':
-      return 'BUY_WIDGET'
-    default: {
-      return rawOrigin
-    }
-  }
-}
-
-const simpleBuyPaymentTypeDictionary = (rawPaymentType: SBPaymentTypes) => {
-  switch (rawPaymentType) {
-    case 'USER_CARD': {
-      return 'PAYMENT_CARD'
-    }
-    case 'LINK_BANK': {
-      return 'BANK_TRANSFER'
-    }
-    default: {
-      return rawPaymentType
-    }
-  }
-}
-
-const getOriginalTimestamp = () => new Date().toISOString()
-
-const id = v4()
-
-const analytics = queuevent<AnalyticsKey, AnalyticsPayload>({
-  queueCallback: async (rawEvents) => {
-    const res = await fetch('/wallet-options-v4.json')
-    const options = await res.json()
-
-    const analyticsURL = `${options.domains.api}/events/publish`
-
-    const nabuId = rawEvents.find((event) => event.payload.id)?.payload.id || null
-
-    const context: { traits: { nabu_id: string | null } } = {
-      traits: {
-        nabu_id: nabuId
-      }
-    }
-
-    const events = rawEvents.map((event) => {
-      const name = event.key
-
-      const { id, originalTimestamp, type, ...properties } = event.payload
-
-      return {
-        id,
-        name,
-        originalTimestamp,
-        properties,
-        type
-      }
-    })
-
-    await fetch(analyticsURL, {
-      body: JSON.stringify({
-        context,
-        events,
-        id
-      }),
-      credentials: 'include',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      method: 'POST'
-    })
-  },
-  queueName: 'analytics'
-})
+import { ModalNamesType } from 'data/types'
 
 const analyticsMiddleware = () => (store) => (next) => (action) => {
   const state = store.getState()
-  const id = state.profile.userData.getOrElse({})?.id
+  const nabuId = state.profile.userData.getOrElse({})?.id
+  const id = state.walletPath.wallet.guid
 
   try {
     switch (action.type) {
@@ -262,6 +32,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.DASHBOARD_CLICKED, {
               id,
+              nabuId,
               origin: 'SIGN_IN',
               originalTimestamp: getOriginalTimestamp(),
               type: AnalyticsType.EVENT
@@ -269,6 +40,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.DASHBOARD_VIEWED, {
               id,
+              nabuId,
               originalTimestamp: getOriginalTimestamp(),
               path: pathname,
               referrer,
@@ -277,6 +49,8 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
               type: AnalyticsType.EVENT,
               url: href
             })
+
+            analytics.clear()
 
             break
           }
@@ -300,6 +74,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.BUY_SELL_CLICKED, {
               id,
+              nabuId,
               origin,
               originalTimestamp: getOriginalTimestamp(),
               type: AnalyticsType.EVENT
@@ -307,6 +82,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.BUY_SELL_VIEWED, {
               id,
+              nabuId,
               originalTimestamp: getOriginalTimestamp(),
               path: pathname,
               referrer,
@@ -325,6 +101,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.SWAP_CLICKED, {
               id,
+              nabuId,
               origin,
               originalTimestamp: getOriginalTimestamp(),
               type: AnalyticsType.EVENT
@@ -332,6 +109,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.SWAP_VIEWED, {
               id,
+              nabuId,
               originalTimestamp: getOriginalTimestamp(),
               path: pathname,
               referrer,
@@ -348,6 +126,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
             analytics.push(AnalyticsKey.SWAP_CLICKED, {
               id,
+              nabuId,
               originalTimestamp: getOriginalTimestamp(),
               tier,
               type: AnalyticsType.EVENT
@@ -374,6 +153,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
           input_amount: inputAmount,
           input_currency: inputCurrency,
           max_card_limit: inputAMountMax,
+          nabuId,
 
           originalTimestamp: getOriginalTimestamp(),
 
@@ -394,6 +174,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
           id,
           input_currency: inputCurrency,
           max_card_limit: maxCardLimit,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           output_currency: outputCurrency,
           platform: 'WALLET',
@@ -409,6 +190,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
         analytics.push(AnalyticsKey.BUY_AMOUNT_MIN_CLICKED, {
           id,
           input_currency: inputCurrency,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           output_currency: outputCurrency,
           platform: 'WALLET',
@@ -421,6 +203,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
 
         analytics.push(AnalyticsKey.BUY_PAYMENT_METHOD_SELECTED, {
           id,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           payment_type: simpleBuyPaymentTypeDictionary(paymentType),
           platform: 'WALLET',
@@ -439,6 +222,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
           card_type: cardType,
           country_billing: countryBilling,
           id,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           product: 'SIMPLE_BUY',
           reason,
@@ -449,6 +233,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
       case AT.auth.VERIFY_EMAIL_TOKEN_SUCCESS: {
         analytics.push(AnalyticsKey.EMAIL_VERIFICATION_REQUESTED, {
           id,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           type: AnalyticsType.EVENT
         })
@@ -457,6 +242,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
       case AT.auth.LOGIN_SUCCESS: {
         analytics.push(AnalyticsKey.SIGNED_IN, {
           id,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           platform: 'WALLET',
           type: AnalyticsType.EVENT
@@ -466,6 +252,7 @@ const analyticsMiddleware = () => (store) => (next) => (action) => {
       case AT.auth.LOGOUT: {
         analytics.push(AnalyticsKey.SIGNED_OUT, {
           id,
+          nabuId,
           originalTimestamp: getOriginalTimestamp(),
           platform: 'WALLET',
           type: AnalyticsType.EVENT
